grib_processing/requirements.txt
# Zależności Python dla przetwarzania GRIB
eccodes
matplotlib
pyyaml
